feat: move to use same calendar component
