+++ frontend/src/App.tsx
@@ -11,13 +11,13 @@ import FeedbackBanner from "./components/common/Banner";
 import Footer from "./components/common/Footer";
 import Header from "./components/common/Header";
 import Account from "./components/pages/Account";
-import CheckInsPage from "./components/pages/CheckInsPage";
 import Dashboard from "./components/pages/Dashboard";
 import EditDashboardSchedulePage from "./components/pages/Dashboard/EditDashboardSchedule";
 import Home from "./components/pages/Home";
 import NotFound from "./components/pages/NotFound";
 import Scheduling from "./components/pages/Scheduling";
 import UserManagement from "./components/pages/UserManagement";
+import CheckInsPage from "./components/pages/ViewCheckInPage";
 import ViewDonationsPage from "./components/pages/ViewDonationsPage";
 import VolunteerShiftsPage from "./components/pages/VolunteerShifts";
 import {
