--- frontend/src/components/pages/CheckInsPage.tsx
@@ -1,14 +1,0 @@
-import { Container } from "@chakra-ui/react";
-import React from "react";
-
-import ViewCheckIns from "./ViewCheckIns";
-
-const CheckInsPage = () => {
-  return (
-    <Container variant="responsiveContainer">
-      <ViewCheckIns />
-    </Container>
-  );
-};
-
-export default CheckInsPage;
